backend/requirements.txt
# FastAPI and Web
fastapi==0.109.0
uvicorn==0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Machine Learning
deepface==0.0.79
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.3.2
joblib==1.3.2

# External APIs
spotipy==2.24.0
requests==2.31.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-dotenv==1.0.0

# Sentiment Analysis
textblob==0.17.1

# Development and Testing
pytest==8.0.0
httpx==0.26.0
